display opt, aykm
the arms of the match in regspec_label referenced tables that were not
const. consequently, they would be rebuilt when reached, every time the
match is incanted. this holds through even when regspec_label is
inlined.

each arm could be a const array for a small and easy change, but to
avoid the indirect dispatch on spec.bank i've reorganized register names
into a single const array and selected values for `RegisterBank` such
that indices into that array can be formed.

for my next trick, i may make `REG_NAMES` a `*const u8`, with indices
  picking offsets into the table - 8-byte offsets might do? this should
  compact down size a little more by removing a pointer and size qword
  for each string.