fix decoding of rex-prefixed modrm+sib operands selecting index 0b100 and base 0b101
for memory operands with a base, index, and displacement either
the wrong base would be selected (register number ignored, so only
`*ax` or `r8*` would be reported), or yaxpeax-x86 would report a
base register is present when it is not (`RegIndexBaseScaleDisp`
when the operand is actually `RegScaleDisp`)

thank you to Evan Johnson for catching and reporting this bug!

also bump crate version to 0.1.4 as this will be immediately tagged and
released.