deprecate `pub fn cs`, which is an intensely embarrassing bug of a function
unlike every other function to test if a particular selector is picked
by prefixes, `Prefixes::cs` does not return bool, nor does it check the
currently-selected prefix. instead, it modifies the decoded `Prefixes`
to set the current prefix to `cs`.

this has been a bug all the way since 0.0.1 was released. the function
now does nothing, and is marked deprecated.

in a future 2.x release, the function will be changed to return `bool`
and be in-line with other segment selector-checking functions. in the
mean time, a new `Prefixes::selects_cs()` does the correct thing.

thank you to @meithecatte who pointed this out in
https://github.com/iximeow/yaxpeax-x86/issues/28!